# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pigeon API",
    "version": "0.1.0",
    "description": "Pigeon Notifications API for sending, listing, streaming, and mutating end-user notifications."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "tags": [
    {
      "name": "Health"
    },
    {
      "name": "Notifications"
    },
    {
      "name": "Auth"
    },
    {
      "name": "Realtime"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Use a Pigeon API key, e.g. pk_test_xxx or pk_live_xxx."
      },
      "UserJwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT minted by POST /v1/users/{userId}/token."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string",
                "nullable": true
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "SendNotificationRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "minLength": 1
          },
          "type": {
            "type": "string",
            "minLength": 1
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "body": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "idempotencyKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          }
        },
        "required": [
          "userId",
          "type",
          "title"
        ]
      },
      "NotificationStatus": {
        "type": "string",
        "enum": [
          "queued",
          "delivered",
          "failed"
        ]
      },
      "SendNotificationResponse": {
        "type": "object",
        "properties": {
          "notificationId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/NotificationStatus"
          }
        },
        "required": [
          "notificationId",
          "status"
        ]
      },
      "CreateUserTokenRequest": {
        "type": "object",
        "properties": {
          "ttlSeconds": {
            "type": "integer",
            "minimum": 1,
            "maximum": 86400
          }
        }
      },
      "CreateUserTokenResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "token",
          "expiresAt"
        ]
      },
      "NotificationItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "status": {
            "$ref": "#/components/schemas/NotificationStatus"
          },
          "readAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "archivedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "type",
          "title",
          "status",
          "createdAt"
        ]
      },
      "NotificationsListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationItem"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "MarkReadResponse": {
        "type": "object",
        "properties": {
          "updated": {
            "type": "boolean"
          },
          "readAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "updated"
        ]
      },
      "MarkAllReadResponse": {
        "type": "object",
        "properties": {
          "updatedCount": {
            "type": "integer",
            "minimum": 0
          },
          "readAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "updatedCount",
          "readAt"
        ]
      },
      "ArchiveResponse": {
        "type": "object",
        "properties": {
          "updated": {
            "type": "boolean"
          },
          "archivedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "updated"
        ]
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service health state",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "service": {
                      "type": "string",
                      "example": "pigeon-api"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "status",
                    "service",
                    "timestamp"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/notifications": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Send notification",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendNotificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Idempotent replay hit existing notification",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendNotificationResponse"
                }
              }
            }
          },
          "201": {
            "description": "Notification created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendNotificationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "List notifications for authenticated end-user",
        "security": [
          {
            "UserJwtAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unread",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated notification list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid/expired JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/{userId}/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create user JWT for frontend SDK",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserTokenRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/notifications/{id}/read": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark notification as read",
        "security": [
          {
            "UserJwtAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Read state updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkReadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/notifications/read-all": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark all unread notifications as read",
        "security": [
          {
            "UserJwtAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Bulk read operation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkAllReadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/notifications/{id}/archive": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Archive notification",
        "security": [
          {
            "UserJwtAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archive result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stream": {
      "get": {
        "tags": [
          "Realtime"
        ],
        "summary": "Subscribe to realtime notification events (SSE)",
        "description": "EventSource stream. Supports replay via Last-Event-ID header when reconnecting.",
        "security": [
          {
            "UserJwtAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "text/event-stream connection",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "example": "id: 1770970817610-0\\nevent: notification.created\\ndata: {\"id\":\"...\",\"type\":\"order.shipped\"}\\n\\n"
                }
              }
            }
          }
        }
      }
    }
  }
}